agg'ing data for player stats in rust

+++ sbs-v1-rust-backend/rust_backend/src/db/nba_games_avgs_historical_mongo_dao.rs
@@ -7,7 +7,7 @@ use mongodb::error::Result;
 pub async fn get_nba_games_avgs_by_team_and_season(
     collection: &Collection<Document>, 
     team_id: f64, 
-    season: f64
+    season: u32
 ) -> Result<Vec<NbaGamesAvgsHistorical>> {
 
     let query = doc! {
