+++ sbs-v1-rust-backend/rust_backend/src/models/services/get_nba_games_by_team_and_season_request.rs
@@ -5,5 +5,5 @@ pub struct GetNbaGamesByTeamAndSeasonRequest {
     #[serde(rename = "teamId")]
     pub team_id: f64,
 
-    pub season: f64,
+    pub season: u32,
 }
