+++ sbs-v1-rust-backend/rust_backend/src/models/services/get_nba_odds_by_team_and_season_request.rs
@@ -6,5 +6,5 @@ pub struct GetNbaOddsByTeamAndSeasonRequest {
     #[serde(rename = "teamName")]
     pub team_name: String,
 
-    pub season: f64,
+    pub season: u32,
 }
