+++ sbs-v1-rust-backend/rust_backend/src/models/services/get_nba_players_by_team_and_season_request.rs
@@ -3,6 +3,6 @@ use serde::Deserialize;
 #[derive(Debug, Deserialize, Clone)]
 pub struct GetNbaPlayersByTeamAndSeasonRequest {
     #[serde(rename = "teamId")]
-    pub team_id: i32,
-    pub season: i32
+    pub team_id: f64,
+    pub season: u32
 }
